Move mock classes
